requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
pillow==10.3.0
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.2
sympy==1.12
tbb==2021.12.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
